# CUDA: mul_mat_vec_q tiling, refactor mul mat logic