This PR does the following:

* Refactor `ggml_cuda_mul_mat` and simplify the logic for choosing between different matrix multiplication kernels. This also fixes P100s not being treated as having good FP16 performance.
* Add tiling in src0 rows to `mul_mat_vec_q`. This increases arithmetic intensity and results in higher t/s for batch sizes > 1. This and a reduction in the number of warps for batch sizes 5-8 (to reduce register pressure) mostly fixes the performance regression sometimes seen when increasing batch sizes. I increased the maximum batch size for `mul_mat_vec_q` to 8. There are still some cases where the performance regresses slightly as you increase the batch size if you get unlucky with occupancy but these cases should be rare now.
* Refactor `mul_mat_vec_q` to use as few registers as possible so that there are more for loop unrolling. This increases performance on average but it again is possible to get unlucky with occupancy so there are also some cases where performance is ~1% worse.
* Removed the option to compile a `mul_mat_vec_q` kernel with variable `ncols_y`. With a batch size of 8 the kernel seems to already be hitting its limits and making `ncols_y` purely a template parameter makes it possible to reduce compilation time.

I'm currently too tired to re-run all of the performance tests; I'll post results tomorrow.

Because of the loop unrolling for the new kernels the compilation time has increased. On master the command

```
make clean && time make LLAMA_CUBLAS=1 LLAMA_NO_CCACHE=1 main
```

reports 13.443 s, with this PR it's 18.691 s. It may make sense to add an option like `LLAMA_FAST_COMPILE` that reduces compilation time as much as possible.