@cebtenzzre 2024-02-09
> reports 13.443 s, with this PR it's 18.691 s. It may make sense to add an option like `LLAMA_FAST_COMPILE` that reduces compilation time as much as possible.

I hate to say it, but this is one of the downsides of splitting a repo into as few files as possible - compilation is very serial and inefficient. cmake is probably worse right now because it seems to be building and archiving static libraries before compiling the examples and tests that depend on them.