@slaren 2024-02-09
In this case the reason of the high compilation time is the number of template instantiations. I am not sure that anything short of putting each kernel in a different source file would help with that.