@slaren 2024-02-09
Some results:

| GPU         | Model         | Test   |   t/s master |   t/s cuda-faster-mmvq-12 |   Speedup |
|:------------|:--------------|:-------|-------------:|--------------------------:|----------:|
| RTX 3090 Ti | llama 7B Q4_0 | tg1    |       132.59 |                    133.54 |      1.01 |
| RTX 3090 Ti | llama 7B Q4_0 | pp2    |       258.57 |                    260.18 |      1.01 |
| RTX 3090 Ti | llama 7B Q4_0 | pp3    |       360.68 |                    381.09 |      1.06 |
| RTX 3090 Ti | llama 7B Q4_0 | pp4    |       429.91 |                    489.20 |      1.14 |
| RTX 3090 Ti | llama 7B Q4_0 | pp5    |       355.36 |                    504.63 |      1.42 |
| RTX 3090 Ti | llama 7B Q4_0 | pp6    |       425.44 |                    577.27 |      1.36 |
| RTX 3090 Ti | llama 7B Q4_0 | pp7    |       490.16 |                    623.50 |      1.27 |
| RTX 3090 Ti | llama 7B Q4_0 | pp8    |       560.78 |                    691.22 |      1.23 |