@ggerganov 2024-02-10
Results on V100

```bash
bash scripts/compare-commits.sh master cuda-faster-mmvq-12 -p 2,3,4,5,6,7,8 -n 1 -r 100 -m $(echo models-mnt/open-llama/7B-v2/ggml-model-*.gguf | sed -e "s/ /,/g")
```

<details>

  Device 0: Tesla V100-PCIE-16GB, compute capability 7.0, VMM: yes

| GPU            | Model           | Test   |   t/s master |   t/s cuda-faster-mmvq-12 |   Speedup |
|:---------------|:----------------|--------|-------------:|--------------------------:|----------:|
| V100-PCIE-16GB | llama 7B F16    | pp2    |        88.54 |                     88.54 |      1.00 |
| V100-PCIE-16GB | llama 7B F16    | pp3    |       131.61 |                    132.17 |      1.00 |
| V100-PCIE-16GB | llama 7B F16    | pp4    |       175.00 |                    175.30 |      1.00 |
| V100-PCIE-16GB | llama 7B F16    | pp5    |       216.35 |                    216.67 |      1.00 |
| V100-PCIE-16GB | llama 7B F16    | pp6    |       251.15 |                    252.34 |      1.00 |
| V100-PCIE-16GB | llama 7B F16    | pp7    |       290.77 |                    291.89 |      1.00 |
| V100-PCIE-16GB | llama 7B F16    | pp8    |       331.40 |                    332.58 |      1.00 |
| V100-PCIE-16GB | llama 7B F16    | tg1    |        51.69 |                     51.70 |      1.00 |
| V100-PCIE-16GB | llama 7B Q2_K_M | pp2    |       164.05 |                    174.92 |      1.07 |
| V100-PCIE-16GB | llama 7B Q2_K_M | pp3    |       206.59 |                    211.91 |      1.03 |
| V100-PCIE-16GB | llama 7B Q2_K_M | pp4    |       232.01 |                    246.09 |      1.06 |
| V100-PCIE-16GB | llama 7B Q2_K_M | pp5    |       170.69 |                    258.84 |      1.52 |
| V100-PCIE-16GB | llama 7B Q2_K_M | pp6    |       205.15 |                    279.95 |      1.36 |
| V100-PCIE-16GB | llama 7B Q2_K_M | pp7    |       238.27 |                    269.67 |      1.13 |
| V100-PCIE-16GB | llama 7B Q2_K_M | pp8    |       270.43 |                    284.50 |      1.05 |
| V100-PCIE-16GB | llama 7B Q2_K_M | tg1    |        97.44 |                     98.48 |      1.01 |
| V100-PCIE-16GB | llama 7B Q3_K_M | pp2    |       155.60 |                    159.43 |      1.02 |
| V100-PCIE-16GB | llama 7B Q3_K_M | pp3    |       200.98 |                    210.37 |      1.05 |
| V100-PCIE-16GB | llama 7B Q3_K_M | pp4    |       238.34 |                    249.48 |      1.05 |
| V100-PCIE-16GB | llama 7B Q3_K_M | pp5    |       141.75 |                    279.57 |      1.97 |
| V100-PCIE-16GB | llama 7B Q3_K_M | pp6    |       169.70 |                    302.38 |      1.78 |
| V100-PCIE-16GB | llama 7B Q3_K_M | pp7    |       196.81 |                    320.81 |      1.63 |
| V100-PCIE-16GB | llama 7B Q3_K_M | pp8    |       222.65 |                    320.36 |      1.44 |
| V100-PCIE-16GB | llama 7B Q3_K_M | tg1    |        87.63 |                     87.35 |      1.00 |
| V100-PCIE-16GB | llama 7B Q4_0   | pp2    |       208.03 |                    222.05 |      1.07 |
| V100-PCIE-16GB | llama 7B Q4_0   | pp3    |       265.53 |                    274.66 |      1.03 |
| V100-PCIE-16GB | llama 7B Q4_0   | pp4    |       308.18 |                    329.83 |      1.07 |
| V100-PCIE-16GB | llama 7B Q4_0   | pp5    |       290.75 |                    394.38 |      1.36 |
| V100-PCIE-16GB | llama 7B Q4_0   | pp6    |       346.87 |                    438.34 |      1.26 |
| V100-PCIE-16GB | llama 7B Q4_0   | pp7    |       399.16 |                    462.65 |      1.16 |
| V100-PCIE-16GB | llama 7B Q4_0   | pp8    |       455.37 |                    488.27 |      1.07 |
| V100-PCIE-16GB | llama 7B Q4_0   | tg1    |       113.33 |                    114.01 |      1.01 |
| V100-PCIE-16GB | llama 7B Q4_1   | pp2    |       214.30 |                    215.34 |      1.00 |
| V100-PCIE-16GB | llama 7B Q4_1   | pp3    |       282.61 |                    268.95 |      0.95 |
| V100-PCIE-16GB | llama 7B Q4_1   | pp4    |       328.24 |                    327.83 |      1.00 |
| V100-PCIE-16GB | llama 7B Q4_1   | pp5    |       307.88 |                    396.97 |      1.29 |
| V100-PCIE-16GB | llama 7B Q4_1   | pp6    |       368.81 |                    447.44 |      1.21 |
| V100-PCIE-16GB | llama 7B Q4_1   | pp7    |       425.87 |                    483.65 |      1.14 |
| V100-PCIE-16GB | llama 7B Q4_1   | pp8    |       483.55 |                    502.72 |      1.04 |
| V100-PCIE-16GB | llama 7B Q4_1   | tg1    |       107.61 |                    108.37 |      1.01 |
| V100-PCIE-16GB | llama 7B Q4_K_M | pp2    |       180.23 |                    188.33 |      1.04 |
| V100-PCIE-16GB | llama 7B Q4_K_M | pp3    |       227.60 |                    234.35 |      1.03 |
| V100-PCIE-16GB | llama 7B Q4_K_M | pp4    |       268.04 |                    267.35 |      1.00 |
| V100-PCIE-16GB | llama 7B Q4_K_M | pp5    |       229.58 |                    303.87 |      1.32 |
| V100-PCIE-16GB | llama 7B Q4_K_M | pp6    |       273.90 |                    316.38 |      1.16 |
| V100-PCIE-16GB | llama 7B Q4_K_M | pp7    |       317.11 |                    328.78 |      1.04 |
| V100-PCIE-16GB | llama 7B Q4_K_M | pp8    |       360.99 |                    306.84 |      0.85 |
| V100-PCIE-16GB | llama 7B Q4_K_M | tg1    |       106.52 |                    106.79 |      1.00 |
| V100-PCIE-16GB | llama 7B Q5_0   | pp2    |       188.48 |                    197.34 |      1.05 |
| V100-PCIE-16GB | llama 7B Q5_0   | pp3    |       250.35 |                    250.74 |      1.00 |
| V100-PCIE-16GB | llama 7B Q5_0   | pp4    |       295.23 |                    303.51 |      1.03 |
| V100-PCIE-16GB | llama 7B Q5_0   | pp5    |       173.57 |                    362.25 |      2.09 |
| V100-PCIE-16GB | llama 7B Q5_0   | pp6    |       207.53 |                    404.44 |      1.95 |
| V100-PCIE-16GB | llama 7B Q5_0   | pp7    |       240.17 |                    433.96 |      1.81 |
| V100-PCIE-16GB | llama 7B Q5_0   | pp8    |       273.25 |                    457.64 |      1.67 |
| V100-PCIE-16GB | llama 7B Q5_0   | tg1    |       101.13 |                    100.32 |      0.99 |
| V100-PCIE-16GB | llama 7B Q5_1   | pp2    |       194.37 |                    193.89 |      1.00 |
| V100-PCIE-16GB | llama 7B Q5_1   | pp3    |       261.26 |                    247.25 |      0.95 |
| V100-PCIE-16GB | llama 7B Q5_1   | pp4    |       316.47 |                    305.25 |      0.96 |
| V100-PCIE-16GB | llama 7B Q5_1   | pp5    |       220.07 |                    367.51 |      1.67 |
| V100-PCIE-16GB | llama 7B Q5_1   | pp6    |       262.35 |                    417.17 |      1.59 |
| V100-PCIE-16GB | llama 7B Q5_1   | pp7    |       303.87 |                    453.54 |      1.49 |
| V100-PCIE-16GB | llama 7B Q5_1   | pp8    |       345.11 |                    475.82 |      1.38 |
| V100-PCIE-16GB | llama 7B Q5_1   | tg1    |        98.03 |                     97.04 |      0.99 |
| V100-PCIE-16GB | llama 7B Q5_K_M | pp2    |       170.46 |                    180.16 |      1.06 |
| V100-PCIE-16GB | llama 7B Q5_K_M | pp3    |       219.73 |                    221.50 |      1.01 |
| V100-PCIE-16GB | llama 7B Q5_K_M | pp4    |       257.19 |                    256.52 |      1.00 |
| V100-PCIE-16GB | llama 7B Q5_K_M | pp5    |       195.42 |                    285.05 |      1.46 |
| V100-PCIE-16GB | llama 7B Q5_K_M | pp6    |       233.21 |                    311.29 |      1.33 |
| V100-PCIE-16GB | llama 7B Q5_K_M | pp7    |       270.92 |                    323.66 |      1.19 |
| V100-PCIE-16GB | llama 7B Q5_K_M | pp8    |       307.21 |                    305.84 |      1.00 |
| V100-PCIE-16GB | llama 7B Q5_K_M | tg1    |        98.42 |                     98.05 |      1.00 |
| V100-PCIE-16GB | llama 7B Q6_K   | pp2    |       158.32 |                    166.40 |      1.05 |
| V100-PCIE-16GB | llama 7B Q6_K   | pp3    |       212.05 |                    202.76 |      0.96 |
| V100-PCIE-16GB | llama 7B Q6_K   | pp4    |       254.27 |                    249.03 |      0.98 |
| V100-PCIE-16GB | llama 7B Q6_K   | pp5    |       183.15 |                    284.54 |      1.55 |
| V100-PCIE-16GB | llama 7B Q6_K   | pp6    |       219.96 |                    312.81 |      1.42 |
| V100-PCIE-16GB | llama 7B Q6_K   | pp7    |       254.55 |                    336.80 |      1.32 |
| V100-PCIE-16GB | llama 7B Q6_K   | pp8    |       289.57 |                    343.74 |      1.19 |
| V100-PCIE-16GB | llama 7B Q6_K   | tg1    |        88.67 |                     85.51 |      0.96 |
| V100-PCIE-16GB | llama 7B Q8_0   | pp2    |       157.63 |                    155.48 |      0.99 |
| V100-PCIE-16GB | llama 7B Q8_0   | pp3    |       220.47 |                    226.89 |      1.03 |
| V100-PCIE-16GB | llama 7B Q8_0   | pp4    |       275.39 |                    232.02 |      0.84 |
| V100-PCIE-16GB | llama 7B Q8_0   | pp5    |       208.74 |                    344.63 |      1.65 |
| V100-PCIE-16GB | llama 7B Q8_0   | pp6    |       249.54 |                    314.36 |      1.26 |
| V100-PCIE-16GB | llama 7B Q8_0   | pp7    |       287.99 |                    346.31 |      1.20 |
| V100-PCIE-16GB | llama 7B Q8_0   | pp8    |       326.89 |                    381.50 |      1.17 |
| V100-PCIE-16GB | llama 7B Q8_0   | tg1    |        80.03 |                     79.59 |      0.99 |


</details>