@Artefact2 2024-02-10
| GPU        | Model            | Test   |   t/s master |   t/s pr |   Speedup |
|:-----------|:-----------------|:-------|-------------:|---------:|----------:|
| RX 6750 XT | llama 13B Q4_K_S | pp2    |        56.35 |    56.21 |      1.00 |
| RX 6750 XT | llama 13B Q4_K_S | pp4    |        79.14 |    68.65 |      0.87 |
| RX 6750 XT | llama 13B Q4_K_S | pp8    |        63.60 |    73.19 |      1.15 |
| RX 6750 XT | llama 13B Q4_K_S | pp16   |       125.82 |   125.73 |      1.00 |
| RX 6750 XT | llama 13B Q4_K_S | pp32   |       246.99 |   247.06 |      1.00 |
| RX 6750 XT | llama 13B Q4_K_S | pp64   |       466.89 |   467.10 |      1.00 |
| RX 6750 XT | llama 13B Q4_K_S | pp128  |       506.03 |   506.19 |      1.00 |